cleanup of match logic 2x
